requirements.txt
selenium==4.15.2
pytest==7.4.3
pytest-html==4.1.1
allure-pytest==2.13.2
PyYAML==6.0.1
Faker==19.3.1
webdriver-manager==4.0.1